Feature: color setting
