+++ .idea/deploymentTargetSelector.xml
@@ -11,6 +11,9 @@
       <SelectionState runConfigName="core.designsys">
         <option name="selectionMode" value="DROPDOWN" />
       </SelectionState>
+      <SelectionState runConfigName="core.designsystem">
+        <option name="selectionMode" value="DROPDOWN" />
+      </SelectionState>
     </selectionStates>
   </component>
 </project>
