+++ app/src/main/java/com/acon/acon/MainActivity.kt
@@ -11,7 +11,8 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.acon.acon.ui.theme.AconTheme
+import com.tunaunnie.designsystem.theme.AconTheme
+
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
