+++ core/designsystem/src/main/java/com/tunaunnie/designsystem/theme/Color.kt
@@ -1,2 +1,62 @@
 package com.tunaunnie.designsystem.theme
 
+import androidx.compose.ui.graphics.Color
+
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
+
+val Purple40 = Color(0xFF6650a4)
+val PurpleGrey40 = Color(0xFF625b71)
+val Pink40 = Color(0xFF7D5260)
+
+//Orange
+val Orange = Color(0xFFFF5402)
+val Orange900 = Color(0xFF770B00)
+val Orange800 = Color(0xFF9A0C00)
+val Orange700 = Color(0xFFBB1A00)
+val Orange600 = Color(0xFFD83100)
+val Orange500 = Color(0xFFF24B00)
+val Orange400 = Color(0xFFFF6928)
+val Orange300 = Color(0xFFFF8950)
+val Orange200 = Color(0xFFFFAA7C)
+val Orange100 = Color(0xFFFFCCAD)
+val Orange50 = Color(0xFFFFEEE3)
+
+
+//Brown
+val Brown = Color(0xFF723400)
+val Brown900 = Color(0xFF592E0F)
+val Brown800 = Color(0xFF743D14)
+val Brown700 = Color(0xFF8E4E1E)
+val Brown600 = Color(0xFFA7602D)
+val Brown500 = Color(0xFFBE7542)
+val Brown400 = Color(0xFFD28B5B)
+val Brown300 = Color(0xFFE2A278)
+val Brown200 = Color(0xFFF0BC9B)
+val Brown100 = Color(0xFFF9D6C1)
+val Brown50 = Color(0xFFFEF1EA)
+
+
+//Green
+val Green = Color(0xFF3DA13F)
+val Green900 = Color(0xFF004407)
+val Green800 = Color(0xFF00590A)
+val Green700 = Color(0xFF006F13)
+val Green600 = Color(0xFF168424)
+val Green500 = Color(0xFF37993A)
+val Green400 = Color(0xFF57AE54)
+val Green300 = Color(0xFF7AC174)
+val Green200 = Color(0xFF9FD498)
+val Green100 = Color(0xFFC5E6BF)
+val Green50 = Color(0xFFEBF7E9)
+
+
+//Gray
+val White = Color(0xFFFFFFFF)
+val Gray01 = Color(0xFFEDEDED)
+val Gray02 = Color(0xFFCBCBCB)
+val Gray03 = Color(0xFF989898)
+val Gray04 = Color(0xFF656565)
+val Gray05 = Color(0xFF323232)
+val Black = Color(0xFF000000)
